requirements.txt
tesserocr
pillow
pypdfium2

paddlepaddle
#paddlepaddle-gpu
paddleocr


azure-ai-documentintelligence


boto3
amazon-textract-response-parser
amazon-textract-prettyprinter



llmwhisperer-client
